apple-toolbox -> "3.1.4"; Fix SharedPackages paths; Add missing files to tests target
