+++ SharedPackages/BrowserServicesKit/Sources/Bookmarks/BookmarkEntity.swift
@@ -15,7 +15,6 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 //
-//
 
 import Foundation
 import CoreData
