+++ SharedPackages/BrowserServicesKit/Sources/Bookmarks/BookmarkletExtensions.swift
@@ -15,7 +15,6 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 //
-//
 
 import Foundation
