--- SharedPackages/BrowserServicesKit/Sources/Common/Extensions/RandomAccessCollectionExtension.swift
@@ -1,32 +1,0 @@
-//
-//  RandomAccessCollectionExtension.swift
-//
-//  Copyright © 2021 DuckDuckGo. All rights reserved.
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//  http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-
-import Foundation
-
-public extension RandomAccessCollection {
-
-    subscript(safe index: Index) -> Element? {
-        guard self.indices.contains(index) else { return nil }
-        return self[index]
-    }
-
-    subscript(_ index: Index, default value: Element) -> Element {
-        return self[safe: index] ?? value
-    }
-
-}
