+++ iOS/LocalPackages/Waitlist/Package.swift
@@ -16,7 +16,7 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/duckduckgo/DesignResourcesKit", exact: "5.0.0"),
-        .package(url: "https://github.com/duckduckgo/apple-toolbox.git", exact: "3.1.3"),
+        .package(url: "https://github.com/duckduckgo/apple-toolbox.git", exact: "3.1.4"),
     ],
     targets: [
         .target(
